app/requirements.txt
mysql-connector
numpy
matplotlib
keras
pandas
scikit-learn
scipy
tensorflow